more fixes for example components and remote runtimes
