+++ rats-examples-datasets/pyproject.toml
@@ -5,7 +5,7 @@ version = "0.1.3"
 readme = "README.md"
 authors = []
 packages = [
-    { include = "rats/exampledatasets", from = "src/python" },
+    { include = "rats/examples", from = "src/python" },
 ]
 repository = "https://github.com/microsoft/rats/"
 documentation = "https://microsoft.github.io/rats/"
@@ -38,7 +38,7 @@ sphinx-markdown-builder = "*"
 [tool.poetry.scripts]
 
 [tool.poetry.plugins."rats.examples"]
-"rats.exampledatasets" = "rats.exampledatasets:PluginContainer"
+"rats.examples" = "rats.examples:PluginContainer"
 
 [build-system]
 requires = ["poetry-core"]
