--- rats-examples-datasets/src/python/rats/exampledatasets/__init__.py
@@ -1,8 +1,0 @@
-"""Some minimal examples."""
-
-from ._plugin import PluginContainer, PluginServices
-
-__all__ = [
-    "PluginContainer",
-    "PluginServices",
-]
