+++ rats-examples-datasets/src/python/rats/examples/__init__.py
@@ -1,1 +1,8 @@
 """Some dataset examples."""
+
+from ._plugin import PluginContainer, PluginServices
+
+__all__ = [
+    "PluginServices",
+    "PluginContainer",
+]
