+++ rats-examples-datasets/src/python/rats/examples/_texts.py
